+++ src/pycat/toolbox/invitro_fluor_ui.py
@@ -32,6 +32,8 @@ try:
 except Exception:
     label_with_circle = lambda t, **k: t
 import numpy as np
+
+from pycat.utils.pixel_size import pixel_size_um_or_default
 import pandas as pd
 import napari
 from napari.utils.notifications import (
@@ -67,7 +69,7 @@ class InVitroFluorUI:
         return self.central_manager.active_data_class.data_repository
 
     def _mpx(self):
-        return float(self._dr().get('microns_per_pixel_sq', 1.0)) ** 0.5
+        return pixel_size_um_or_default(self._dr(), context='invitro_fluor_ui')
 
     @staticmethod
     def _layer_is_time_series(layer):